+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/LightSpell.java
@@ -1,11 +1,7 @@
 package com.elmakers.mine.bukkit.plugins.magic.spells;
 
-import org.bukkit.Location;
-import org.bukkit.block.Block;
-
 import com.elmakers.mine.bukkit.plugins.magic.Spell;
 import com.elmakers.mine.bukkit.plugins.magic.SpellResult;
-import com.elmakers.mine.bukkit.utilities.LightSource;
 import com.elmakers.mine.bukkit.utilities.borrowed.ConfigurationNode;
 
 public class LightSpell extends Spell {
@@ -13,6 +9,10 @@ public class LightSpell extends Spell {
 	@Override
 	public SpellResult onCast(ConfigurationNode parameters) 
 	{
+		sendMessage("This spell has been disabled for safety!");
+		return SpellResult.FAIL;
+		
+		/*
 		Block target = getTargetBlock();
 		Location targetLocation = getLocation();
 		if (target == null) {
@@ -24,5 +24,6 @@ public class LightSpell extends Spell {
 		LightSource.createLightSource(targetLocation, lightLevel);
 		controller.updateBlock(target);
 		return SpellResult.CAST;
+		*/
 	}
 }