+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.9.6
 
+ - Removed/Disabled the "Light" spell (for now.. ?)
  - Add "Wolf House" and "Tent" and "Cage" spells. WIP.
  - Support different item types for wand templates.
  - Copy command blocks from schematics.